refactor: 문자열을 숫자열로 변환 및 유효성 체크 기능 분리
