+++ src/test/java/calculator/CalculatorTest.java
@@ -4,7 +4,6 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 public class CalculatorTest {
     private Calculator calculator;
@@ -14,20 +13,6 @@ public class CalculatorTest {
     public void checkOneSource() {
         calculator = new Calculator(new String[]{"1"});
         assertThat(calculator.sum()).isEqualTo(1);
-    }
-
-    @Test
-    @DisplayName("숫자가 아닌 값이 배열에 있을 경우")
-    public void checkNotNumber() {
-        String[] target = {"1", "a", "b"};
-        assertThatThrownBy(() -> calculator = new Calculator(target)).isInstanceOf(RuntimeException.class);
-    }
-
-    @Test
-    @DisplayName("음수가 배열에 있을 경우")
-    public void checkNegativeNumber() {
-        String[] target = {"1", "-3", "-5"};
-        assertThatThrownBy(() -> calculator = new Calculator(target)).isInstanceOf(RuntimeException.class);
     }
 
     @Test
